# docs/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GOF API",
        "description": "API REST do GOF (Gestão Operacional de Frotas). Todas as respostas, inclusive erros, usam o envelope {status, message, errors, data[, meta]}.",
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 (mesma origem, via nginx)"
        }
    ],
    "paths": {
        "/audit-logs": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Lista o audit trail (somente leitura)",
                "description": "Permissão: audit.view (A). Ordenação fixa: event_at desc, id desc (`sort` é ignorado). `from`/`to` em ISO 8601 sobre event_at, inclusivos; data sem hora = dia inteiro em UTC (`from` às 00:00, `to` até 23:59:59.999999); `from` > `to` → 422. Não há escrita: POST/PUT/PATCH/DELETE → 405.",
                "operationId": "auditLogsIndex",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "name": "auditable_type",
                        "in": "query",
                        "description": "Alias curto do registro auditado; desconhecido → 422.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "branch",
                                "cost_center",
                                "employee",
                                "equipment_family",
                                "user"
                            ]
                        }
                    },
                    {
                        "name": "auditable_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "actor_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created",
                                "updated",
                                "deleted",
                                "restored",
                                "login_succeeded",
                                "login_failed",
                                "logout",
                                "password_changed",
                                "password_reset_requested",
                                "password_reset",
                                "setting_changed",
                                "setting_removed"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "ISO 8601 (date-time ou só a data, dia inteiro em UTC); inclusivo.",
                        "schema": {
                            "type": "string",
                            "example": "2026-10-01T00:00:00Z"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "ISO 8601 (date-time ou só a data, até 23:59:59.999999 UTC); inclusivo; menor que `from` → 422.",
                        "schema": {
                            "type": "string",
                            "example": "2026-10-31"
                        }
                    },
                    {
                        "name": "request_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Envelope"
                                        },
                                        {
                                            "required": [
                                                "meta"
                                            ],
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AuditLog"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/audit-logs/{audit_log}": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Detalha um evento do audit trail",
                "description": "Permissão: audit.view (A). Inexistente → 404.",
                "operationId": "auditLogsShow",
                "parameters": [
                    {
                        "name": "audit_log",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuditLogEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuditLogEnvelope": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Envelope"
                    },
                    {
                        "properties": {
                            "data": {
                                "$ref": "#/components/schemas/AuditLog"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "AuditLog": {
                "required": [
                    "id",
                    "uuid",
                    "event_at",
                    "actor",
                    "action",
                    "auditable",
                    "old_values",
                    "new_values",
                    "metadata",
                    "ip",
                    "request_id"
                ],
                "properties": {
                    "id": {
                        "description": "Item de /audit-logs no formato da F.1. `actor` vem do snapshot gravado no evento (nome/perfil na\népoca), null em ações do sistema; `auditable.type` é o alias curto (AuditLog::AUDITABLE_TYPES).",
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "event_at": {
                        "description": "UTC com microssegundos.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-08T14:03:22.123456Z"
                    },
                    "actor": {
                        "description": "Quem executou (snapshot da época); null = sistema ou login_failed.",
                        "required": [
                            "id",
                            "name",
                            "role"
                        ],
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Ana Souza",
                                "nullable": true
                            },
                            "role": {
                                "type": "string",
                                "nullable": true,
                                "enum": [
                                    "operator",
                                    "mechanic",
                                    "leader",
                                    "admin"
                                ]
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "action": {
                        "type": "string",
                        "enum": [
                            "created",
                            "updated",
                            "deleted",
                            "restored",
                            "login_succeeded",
                            "login_failed",
                            "logout",
                            "password_changed",
                            "password_reset_requested",
                            "password_reset",
                            "setting_changed",
                            "setting_removed"
                        ]
                    },
                    "auditable": {
                        "description": "Registro afetado; null em eventos sem alvo (ex.: login_failed).",
                        "required": [
                            "type",
                            "id"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "cost_center"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1,
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "old_values": {
                        "description": "Valores anteriores (só campos alterados em updated); sem dados sensíveis.",
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "new_values": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "ip": {
                        "type": "string",
                        "example": "172.18.0.1",
                        "nullable": true
                    },
                    "request_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Envelope": {
                "description": "Envelope padrão de toda resposta.",
                "required": [
                    "status",
                    "message",
                    "errors",
                    "data"
                ],
                "properties": {
                    "status": {
                        "description": "Definições globais da OpenAPI (info, servidor, segurança, schemas e respostas reutilizáveis).\nOs endpoints são anotados nos próprios controllers. O atributo não aceita expressão, então\ntítulo e descrição usam os marcadores de OpenApiBuilder::BRAND_TOKENS, preenchidos com a\nmarca do config (app.name / app.full_name) na geração.",
                        "type": "string",
                        "enum": [
                            "success",
                            "error"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "data": {
                        "description": "Objeto, lista ou null conforme o endpoint (qualquer tipo)."
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "description": "Presente (`meta`) apenas em listas paginadas.",
                "required": [
                    "current_page",
                    "per_page",
                    "total",
                    "last_page"
                ],
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 0
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "ErrorEnvelope": {
                "description": "Resposta de erro (401, 403, 404, 405, 409, 429, 500, 503).",
                "required": [
                    "status",
                    "message",
                    "errors",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "error"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "example": "Recurso não encontrado."
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "data": {
                        "description": "Sempre null em erros.",
                        "example": null
                    }
                },
                "type": "object"
            },
            "ValidationErrorEnvelope": {
                "description": "Erro de validação (422): `errors` mapeia campo → mensagens.",
                "required": [
                    "status",
                    "message",
                    "errors",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "error"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "example": "Os dados informados são inválidos."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "O campo email é obrigatório."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "data": {
                        "description": "Sempre null em erros.",
                        "example": null
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthenticated": {
                "description": "Não autenticado (401).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Acesso negado (403).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Recurso não encontrado (404).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                }
            },
...
            "ValidationError": {
                "description": "Dados inválidos (422).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorEnvelope"
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "Limite de requisições excedido (429).",
                "headers": {
                    "Retry-After": {
                        "description": "Segundos até poder tentar novamente.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorEnvelope"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Página (1..n).",
                "schema": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                }
            },
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Itens por página: máx. 100; até 200 com `is_active=1` (selects).",
                "schema": {
                    "type": "integer",
                    "default": 15,
                    "maximum": 200,
                    "minimum": 1
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /auth/login (expira em 12 h).",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}